+++ bin/foreign-key-list-fetcher.php
@@ -11,35 +11,19 @@
 require_once dirname(__DIR__).'/vendor/autoload.php';
 $baseDir = dirname(__DIR__);
 
-use SURFnet\VPN\Common\CliParser;
 use SURFnet\VPN\Common\Config;
 use SURFnet\VPN\Portal\ForeignKeyListFetcher;
 use SURFnet\VPN\Portal\HttpClient\CurlHttpClient;
 
 try {
-    $p = new CliParser(
-        'Fetch foreign key list.',
-        [
-            'instance' => ['the VPN instance', true, false],
-        ]
-    );
-
-    $opt = $p->parse($argv);
-    if ($opt->hasItem('help')) {
-        echo $p->help();
-        exit(0);
-    }
-
-    $instanceId = $opt->hasItem('instance') ? $opt->getItem('instance') : 'default';
-
-    $configFile = sprintf('%s/config/%s/config.php', $baseDir, $instanceId);
+    $configFile = sprintf('%s/config/config.php', $baseDir);
     $config = Config::fromFile($configFile);
 
     if ($config->getSection('Api')->hasItem('foreignKeyListSource')) {
         $publicKeysSource = $config->getSection('Api')->getItem('foreignKeyListSource');
         $publicKeysSourcePublicKey = $config->getSection('Api')->getItem('foreignKeyListPublicKey');
 
-        $foreignKeyListFetcher = new ForeignKeyListFetcher(sprintf('%s/data/%s/foreign_key_list.json', $baseDir, $instanceId));
+        $foreignKeyListFetcher = new ForeignKeyListFetcher(sprintf('%s/data/foreign_key_list.json', $baseDir));
         $foreignKeyListFetcher->update(new CurlHttpClient(), $publicKeysSource, $publicKeysSourcePublicKey);
     }
 } catch (Exception $e) {