+++ CHANGES.md
@@ -22,6 +22,7 @@
 - remove Twig, switch to Tpl
 - remove compression framing support
 - remove tls-auth support
+- remove "multi instance" support
 
 ## 1.8.5 (2018-11-28)
 - no longer allow clients to obtain new access tokens using the refresh token